+++ config/ig-feedback.php
@@ -38,6 +38,15 @@ return [
             'label_translation_key' => 'ig-feedback::fields.phone',
             'autocomplete' => 'tel',
         ],
+        'subscribe' => [
+            'type' => 'checkbox',
+            'validation' => 'boolean',
+            'label_translation_key' => 'ig-feedback::fields.subscribe',
+            'value_translation_key' => [
+                1 => 'ig-feedback::fields.subscribe_interested',
+                0 => 'ig-feedback::fields.subscribe_not_interested',
+            ],
+        ],
     ],
     'exclude_attributes' => [
         'name',
@@ -46,6 +55,8 @@ return [
         'validation',
         'label_translation_key',
         'error_translation_key',
+        'value_translation_key',
         'input_view',
+        'values',
     ],
 ];