+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Prepend a section to email body with form name and description.
+- Expose field key in public fields.
+- Add subscribe checkbox field type.
+- Support custom fallback value for fields, default is `n/a`.
+- Support value translation for fields (e.g. boolean to string).
+
 ## [6.0.2] - 2025-12-02
 
 ### Fixed
@@ -349,6 +359,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[Unreleased]: https://https://github.com/internetguru/laravel-feedback/compare/staging...dev
 [6.0.2]: https://https://github.com/internetguru/laravel-feedback/compare/v6.0.1...v6.0.2
 [6.0.1]: https://https://github.com/internetguru/laravel-feedback/compare/v6.0.0...v6.0.1
 [6.0.0]: https://https://github.com/internetguru/laravel-feedback/compare/v5.0.6...v6.0.0