+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
 ## [6.1.0-rc.1] - 2025-12-03
 
 ### Added
@@ -359,6 +361,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[Unreleased]: https://https://github.com/internetguru/laravel-feedback/compare/staging...dev
 [6.1.0-rc.1]: https://github.com/internetguru/laravel-feedback/releases/tag/v6.0.2
 [6.0.2]: https://https://github.com/internetguru/laravel-feedback/compare/v6.0.1...v6.0.2
 [6.0.1]: https://https://github.com/internetguru/laravel-feedback/compare/v6.0.0...v6.0.1